Changed Stats Done with Reset?

+++ GroupProject/src/StatsInterface.java
@@ -4,6 +4,6 @@
  */
 
 public interface StatsInterface {
-	int getCurrentBalance();
+	double getCurrentBalance();
 	int getNumResets();
 }
